+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Oracle driver for Yii Database Change Log
 
+## 1.2.1 under development
+
+- no changes in this release.
+
 ## 1.2.0 November 12, 2023
 
 - Enh #230: Improve column type #230 (@Tigrov)